requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
filelock==3.16.1
Flask==3.0.3
fsspec==2024.9.0
future==1.0.0
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-tools==1.67.0
huggingface-hub==0.26.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
Markdown==3.7
MarkupSafe==3.0.1
mpmath==1.3.0
networkx==3.4.1
nillion-aivm==0.1.1
numpy==1.26.4
omegaconf==2.3.0
onnx==1.17.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.2
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.7.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.2.0
six==1.16.0
slack_bolt==1.21.0
slack_sdk==3.33.1
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
visdom==0.2.4
websocket-client==1.8.0
Werkzeug==3.0.4
